모델 추출하기
